Moved Core analysis types into separate namespace
