+++ src/Core/Analysis/IAnalyzerOptions.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarLint.VisualStudio.Core
+namespace SonarLint.VisualStudio.Core.Analysis
 {
     /// <summary>
     /// Marker interface for any analyzer-specific configuration options
