+++ src/Core/Analysis/IIssueConsumer.cs
@@ -20,7 +20,7 @@
 
 using System.Collections.Generic;
 
-namespace SonarLint.VisualStudio.Core
+namespace SonarLint.VisualStudio.Core.Analysis
 {
     public interface IIssueConsumer
     {
