+++ src/Integration.Vsix.UnitTests/Analysis/AnalyzerControllerTests.cs
@@ -22,11 +22,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
-using EnvDTE;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
-using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Analysis;
 using SonarLint.VisualStudio.Integration.UnitTests;
 
 namespace SonarLint.VisualStudio.Integration.Vsix.Analysis.UnitTests
