+++ src/Integration.Vsix.UnitTests/ErrorList/IssuesSnapshotTests.cs
@@ -27,6 +27,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.VisualStudio.Text;
 using Moq;
 using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Analysis;
 using SonarLint.VisualStudio.Integration.Vsix;
 using DaemonSeverity = Sonarlint.Issue.Types.Severity;
