+++ src/Integration.Vsix/CFamily/CFamilyReproducerCommand.cs
@@ -26,6 +26,7 @@ using Microsoft.VisualStudio.Editor;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Analysis;
 using SonarLint.VisualStudio.Core.CFamily;
 using Task = System.Threading.Tasks.Task;
