+++ src/Integration.Vsix/Helpers/IssueConverter.cs
@@ -20,7 +20,7 @@
 
 using System;
 using Microsoft.VisualStudio.Text;
-using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Analysis;
 
 namespace SonarLint.VisualStudio.Integration.Vsix.Helpers
 {
