+++ src/Integration.Vsix/SonarLintTagger/ISonarLanguageRecognizer.cs
@@ -20,7 +20,7 @@
 
 using System.Collections.Generic;
 using Microsoft.VisualStudio.Text;
-using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Analysis;
 
 namespace SonarLint.VisualStudio.Integration.Vsix
 {
