+++ src/Integration.Vsix/SonarLintTagger/IssueMarker.cs
@@ -19,7 +19,7 @@
  */
 
 using Microsoft.VisualStudio.Text;
-using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Analysis;
 
 namespace SonarLint.VisualStudio.Integration.Vsix
 {
