+++ src/Integration.Vsix/SonarLintTagger/SonarLanguageRecognizer.cs
@@ -25,7 +25,7 @@ using System.IO;
 using System.Linq;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Utilities;
-using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Analysis;
 
 namespace SonarLint.VisualStudio.Integration.Vsix
 {
